Replace hard-coded lambda execution role ARN with locally provisioned one
This is an attempt to solve a minor bug with ARNs being hard coded in a way that assumes a consistent naming scheme for IAM roles provisioned in other projects. This commit attempts to remove this dependency altogether by provisioning an IAM role with the minimum viable set of permissions for its purpose, which only consists of making an HTTP request to the Slack webhooks API. We could also consider passing in ARNs to resource dependencies in other projects as arguments to modules like this, but in this case the dependency may be avoidable altogether due to how non-specific its actual requirements are.