Attempt to fix invalid lambda function handler value for studio infrastructure project
At the time of writing, the behaviour of our lambda module is such the provided function name determines the function handler setting, so providing a function name that includes the correct client value (e.g. moar-send-slack-notifications-studio for Studio) will result in a mismatch with the filename that this commit changes (which was previously moar-send-slack-notifications-authentication, regardless of the value of var.client). I suspect this commit will fail due to using a lambda function name that will clash between usages of this module, and in that case I will revert this commit and try another solution.