Commit Dockerfile for minimal client project build image
This is a smaller image than we've used previously, but is sufficient to run Webpack and send slack notifications with curl. Although this Terraform project can't do anything with the image automatically, it's committed here as a logical place to make the image available in case it needs to be modified and/or reuploaded.